% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.ccv.R
\name{bw.ccv}
\alias{bw.ccv}
\title{Title}
\usage{
bw.ccv(x, lower = 0, upper = 60, tol = 0.1)
}
\arguments{
\item{x}{Data from which the smoothing parameter is to be computed. The object is coerced to class circular.}

\item{lower}{Lower boundary of the interval to be used in the search for the value of the smoothing parameter. Default value lower=0.}

\item{upper}{Upper boundary of the interval to be used in the search for the value of the smoothing parameter. Default value upper=60.}

\item{tol}{Convergence tolerance for optimize.}
}
\value{
something
}
\description{
Title
}
\examples{
something
}
