% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.lscvg.R
\name{bw.lscvg}
\alias{bw.lscvg}
\title{Title}
\usage{
bw.lscvg(x, g = 4, lower = 0, upper = 60, tol = 0.1)
}
\arguments{
\item{x}{Data from which the smoothing parameter is to be computed. The object is coerced to class circular.}

\item{g}{parameter for bigger variability}

\item{lower}{Lower boundary of the interval to be used in the search for the value of the smoothing parameter. Default value lower=0.}

\item{upper}{Upper boundary of the interval to be used in the search for the value of the smoothing parameter. Default value upper=60.}

\item{tol}{Convergence tolerance for optimize.}
}
\value{
something
}
\description{
Title
}
\examples{
something
}
