% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.scv.R
\name{bw.scv}
\alias{bw.scv}
\title{Compute the Optimal Bandwidth for Circular Data using Smoothed Cross-Validation}
\usage{
bw.scv(x, np = 500, lower = 0, upper = 60, tol = 0.1)
}
\arguments{
\item{x}{Data from which the smoothing parameter is to be computed. The object is
coerced to a numeric vector in radians using \code{circular::conversion.circular}.
Can be a numeric vector or an object of class \code{circular}.}

\item{np}{An integer specifying the number of points used in numerical integration
to evaluate the SCV criterion. A higher number increases precision but also
computational cost. Default is 500.}

\item{lower}{Lower boundary of the interval for the optimization of the smoothing
parameter \code{kappa}. Must be a positive numeric value smaller than \code{upper}.
Default is 0.}

\item{upper}{Upper boundary of the interval for the optimization of the smoothing
parameter \code{kappa}. Must be a positive numeric value greater than \code{lower}.
Default is 60.}

\item{tol}{Convergence tolerance used in the \code{optimize} function. Determines how
precisely the optimal value is estimated. Default is 0.1.}
}
\value{
The computed optimal smoothing parameter \code{kappa}, the numeric value
that minimizes the smoothed cross-validation criterion.
}
\description{
This function computes the optimal smoothing parameter (bandwidth) for circular data
using a smoothed cross-validation (SCV) method (see \url{doi:doi.org/10.1007/s00180-023-01401-0}).
It searches for the value of the smoothing parameter \code{kappa} that minimizes the SCV criterion within the
specified interval \verb{[lower, upper]}.
}
\examples{
# Example with circular data
library(circular)
x <- rwrappednormal(100, mu = circular(2), rho = 0.5)
bw <- bw.scv(x)
print(bw)
x <- rvonmises(100, mu = circular(0.5), kappa = 2)
bw <- bw.scv(x)
print(bw)

}
\references{
Zámeček, S., Horová, I., Katina, S., & Hasilová, K. (2023). An adaptive
method for bandwidth selection in circular kernel density estimation.
\emph{Computational Statistics}.
\doi{10.1007/s00180-023-01401-0}
}
