% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.vm.R
\name{bw.vm}
\alias{bw.vm}
\title{Compute the Optimal Bandwidth for Circular Data using von Mises Method}
\usage{
bw.vm(x)
}
\arguments{
\item{x}{Data from which the smoothing parameter is to be computed. The object is
coerced to a numeric vector in radians using \code{circular::conversion.circular}.
Can be a numeric vector or an object of class \code{circular}.}
}
\value{
The computed optimal smoothing parameter, a numeric value derived from
the von Mises distribution fit to the data.
}
\description{
This function computes the optimal smoothing parameter (bandwidth) for circular data
using the von Mises method. The optimal bandwidth is derived based on the maximum
likelihood estimate of the concentration parameter kappa from a von Mises distribution
fitted to the data.
}
\examples{
# Example with circular data
library(circular)
set.seed(123)
x <- rvonmises(100, mu = circular(0), kappa = 2)
bw <- bw.vm(x)
print(bw)

x <- rwrappednormal(100, mu = circular(1), rho = 0.7)
bw <- bw.vm(x)
print(bw)

}
\references{
Tsuruta, Yasuhito & Sagae, Masahiko (2017). Higher order kernel density
estimation on the circle. \emph{Statistics & Probability Letters}, 131:46--50.
\doi{10.1016/j.spl.2017.07.027}
}
