% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.adaptive.circular.R
\name{density.adaptive.circular}
\alias{density.adaptive.circular}
\title{Title}
\usage{
\method{density}{adaptive.circular}(
  x,
  bw0,
  alpha = 0.5,
  type = "n",
  z = NULL,
  from = circular(0),
  to = circular(2 * pi),
  n = 500
)
}
\arguments{
\item{x}{something}

\item{bw0}{something}

\item{alpha}{something}

\item{type}{something}

\item{z}{something}

\item{from}{something}

\item{to}{something}

\item{n}{something}
}
\value{
something
}
\description{
Title
}
\examples{
something
}
