% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.ccv.R
\name{bw.ccv}
\alias{bw.ccv}
\title{Compute the Optimal Bandwidth for Circular Data using Complete Cross-Validation}
\usage{
bw.ccv(x, lower = 0, upper = 60, tol = 0.1)
}
\arguments{
\item{x}{Data from which the smoothing parameter is to be computed. The object is
coerced to a numeric vector in radians using \code{circular::conversion.circular}.
Can be a numeric vector or an object of class \code{circular}.}

\item{lower}{Lower boundary of the interval to be used in the search for the
smoothing parameter \code{nu}. Must be a positive numeric value less than \code{upper}.
Default is 0.}

\item{upper}{Upper boundary of the interval to be used in the search for the
smoothing parameter \code{nu}. Must be a positive numeric value greater than \code{lower}.
Default is 60.}

\item{tol}{Convergence tolerance for the \code{optimize} function, determining the
precision of the optimization process. Default is 0.1.}
}
\value{
The computed optimal smoothing parameter \code{nu}, a numeric value that
minimizes the complete cross-validation criterion within the interval
\verb{[lower, upper]}.
}
\description{
This function calculates the optimal smoothing parameter (bandwidth) for circular data
using the complete cross-validation (CCV) method (see \url{doi:10.59170/stattrans-2024-024}). It searches for the value of the
smoothing parameter \code{nu} that minimizes the CCV criterion within a specified interval
\verb{[lower, upper]}.
}
\examples{
# Example with numeric data in radians
library(circular)
set.seed(123)
x <- rwrappednormal(100, mu = circular(2), rho = 0.5)
bw <- bw.ccv(x)
print(bw)

# Example with circular data
x <- rvonmises(100, mu = circular(0), kappa = 1)
bw <- bw.ccv(x)
print(bw)

}
