% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.jf.R
\name{bw.jf}
\alias{bw.jf}
\title{Compute the Optimal Bandwidth for Circular Data using Jones-Faddy Method}
\usage{
bw.jf(x)
}
\arguments{
\item{x}{Data from which the smoothing parameter is to be computed. The object is
coerced to a numeric vector in radians using \code{circular::conversion.circular}.
Can be a numeric vector or an object of class \code{circular}.}
}
\value{
The computed optimal smoothing parameter, a numeric value derived from
the Jones-Faddy method for circular kernel density estimation.
}
\description{
This function computes the optimal smoothing parameter (bandwidth) for circular data
using the Jones-Faddy method. The method is based on higher-order approximations
to estimate the optimal bandwidth for kernel density estimation with circular data.
}
\examples{
# Example with circular data
library(circular)
set.seed(123)
x <- rvonmises(100, mu = circular(0), kappa = 2)
bw <- bw.jf(x)
print(bw)

x <- rwrappednormal(100, mu = circular(1), rho = 0.7)
bw <- bw.jf(x)
print(bw)

}
\references{
Tsuruta, Yasuhito & Sagae, Masahiko (2017). Higher order kernel density
estimation on the circle. \emph{Statistics & Probability Letters}, 131:46--50.
\doi{10.1016/j.spl.2017.07.027}
}
