% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.lscvg.R
\name{bw.lscvg}
\alias{bw.lscvg}
\title{Compute the Optimal Bandwidth for Circular Data using Generalized Least Squares Cross-Validation}
\usage{
bw.lscvg(x, g = 4, lower = 0, upper = 60, tol = 0.1)
}
\arguments{
\item{x}{Data from which the smoothing parameter is to be computed. The object is
coerced to a numeric vector in radians using \code{circular::conversion.circular}.
Can be a numeric vector or an object of class \code{circular}.}

\item{g}{A numeric scalar that controls the variability in the cross-validation
procedure. It influences the scaling in the internal calculations, affecting the
bandwidth estimation. Default is 4.}

\item{lower}{Lower boundary of the interval to be used in the search for the
smoothing parameter \code{kappa}. Must be a positive numeric value less than \code{upper}.
Default is 0.}

\item{upper}{Upper boundary of the interval to be used in the search for the
smoothing parameter \code{kappa}. Must be a positive numeric value greater than \code{lower}.
Default is 60.}

\item{tol}{Convergence tolerance for the \code{optimize} function, determining the
precision of the optimization process. Default is 0.1.}
}
\value{
The computed optimal smoothing parameter \code{kappa}, a numeric value that
minimizes the least squares cross-validation criterion within the interval
\verb{[lower, upper]}.
}
\description{
This function computes the optimal smoothing parameter (bandwidth) for circular data
using a generalized least squares cross-validation method (see \url{doi:10.1007/s00180-023-01401-0}).
It searches for the value of the smoothing parameter \code{kappa} that minimizes the cross-validation criterion within the
specified interval \verb{[lower, upper]}.
}
\examples{
# Example with circular data
library(circular)
set.seed(123)
x <- rwrappednormal(100, mu = circular(2), rho = 0.5)
bw <- bw.lscvg(x)
print(bw)

x <- rvonmises(100, mu = circular(0), kappa = 1)
bw <- bw.lscvg(x)
print(bw)

}
