% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.ts.R
\name{bw.ts}
\alias{bw.ts}
\title{Compute the Optimal Bandwidth for Circular Data using circular version of multiplicative method from Terrell and Scott.}
\usage{
bw.ts(x)
}
\arguments{
\item{x}{Data from which the smoothing parameter is to be computed. The object is
coerced to a numeric vector in radians using \code{circular::conversion.circular}.
Can be a numeric vector or an object of class \code{circular}.}
}
\value{
The computed optimal smoothing parameter, a numeric value derived from
the circular version of the multiplicative method for circular kernel density estimation.
}
\description{
This function computes the optimal smoothing parameter (bandwidth) for circular data
using the circular version of the multiplicative method from Terrell and Scott (1980). The method
is based on the new class of higher order kernel functions with new moments introduced by Tsurunga (see \url{doi:doi.org/10.1016/j.spl.2017.08.003}).
}
\examples{
# Example with circular data
library(circular)
set.seed(123)
x <- rvonmises(100, mu = circular(0), kappa = 2)
bw <- bw.ts(x)
print(bw)

x <- rwrappednormal(100, mu = circular(1), rho = 0.7)
bw <- bw.ts(x)
print(bw)

}
\references{
Tsuruta, Yasuhito & Sagae, Masahiko (2017). Higher order kernel density
estimation on the circle. \emph{Statistics & Probability Letters}, 131:46--50.
\doi{10.1016/j.spl.2017.07.027}
}
