% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularKDE-package.R
\docType{package}
\name{circularKDE-package}
\alias{circularKDE-package}
\alias{circularKDE}
\title{circularKDE: Recent Methods for Kernel Density Estimation of Circular Data}
\description{
Provides recent kernel density estimation methods for circular data, including adaptive and higher-order techniques. The implementation is based on recent advances in bandwidth selection and circular smoothing. Key methods include adaptive bandwidth selection methods (Zámečník et al. 2024), complete cross-validation (Hasilová et al. 2024), Fourier-based plug-in rules (Tenreiro 2022), and higher-order kernels (Tsuruta & Sagae 2017).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stazam/circularKDE}
  \item Report bugs at \url{https://github.com/stazam/circularKDE/issues}
}

}
\author{
\strong{Maintainer}: Stanislav Zamecnik \email{zamecnik@math.muni.cz}

Authors:
\itemize{
  \item Ivanka Hórová \email{horova@math.muni.cz}
  \item Kamila Hasilová \email{kamila.hasilova@unob.cz}
  \item Stanislav Katina \email{katina@math.muni.cz}
}

}
\keyword{package}
